chore: re-renders optimization
